refactor : 정책 패키지 추가 및 Class 이동
