--- src/main/java/nextstep/subway/path/domain/FareRuleStrategy.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.domain;
-
-public interface FareRuleStrategy {
-
-    int calculateFare(int distance);
-
-}
