+++ src/main/java/nextstep/subway/path/domain/policy/line/LineFareDefaultPolicy.java
@@ -1,8 +1,8 @@
-package nextstep.subway.path.domain;
+package nextstep.subway.path.domain.policy.line;
 
 import static nextstep.subway.path.domain.Fare.DEFAULT_FARE;
 
-public class FareDefaultRuleStrategy implements FareRuleStrategy {
+public class LineFareDefaultPolicy implements LineFarePolicy {
 
     @Override
     public int calculateFare(int distance) {
