+++ src/main/java/nextstep/subway/path/domain/policy/line/LineFarePolicy.java
@@ -1,0 +1,7 @@
+package nextstep.subway.path.domain.policy.line;
+
+public interface LineFarePolicy {
+
+    int calculateFare(int distance);
+
+}
