+++ src/test/java/nextstep/subway/path/domain/FareTest.java
@@ -1,5 +1,7 @@
 package nextstep.subway.path.domain;
 
+import nextstep.subway.path.domain.policy.line.LineFarePolicy;
+import nextstep.subway.path.domain.policy.line.LineFarePolicyFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -12,7 +14,7 @@ class FareTest {
     @CsvSource(value = {"10:1250", "20:1450", "50:2050", "51:2150", "66:2250"}, delimiter = ':')
     @DisplayName("요금 계산")
     void getFare(int distance, int expected) {
-        FareRuleStrategy strategy = FareStrategyFactory.from(distance);
+        LineFarePolicy strategy = LineFarePolicyFactory.from(distance);
         Fare fare = new Fare(strategy, distance);
         assertThat(fare.getFare()).isEqualTo(expected);
     }
